kafka/error: define is_error_code_enum trait in errors.h
* move the is_error_code_enum<kafka::error_code> trait into error.h
  where kafka::error_code is defined. this would help us ease the
  pain of fighting with the failures attached at the end of the
  the commit message.
* move make_error_code() also into errors.{h,cc}.
  because as make_error_code(error_code) is also a part of the error
  handling facility built around error_code, it is "closer" to
  error_code. so let's put them together, for better readability and
  maintainabilty.
* in order to get access to
  `pandaproxy::make_error_condition(pandaproxy::reply_error_code)`,
  from `v::kafka_protocol`, the linkage to `v/pandaproxy/error.cc`
  is added to `v::kafka_protocol` via `v::pandaproxy_common`

----

FAILED: src/v/pandaproxy/CMakeFiles/v_pandaproxy_common.dir/server.cc.o
ccache /usr/bin/clang++-15 -DBOOST_ASIO_HAS_STD_INVOKE_RESULT -DFMT_LOCALE -DFMT_SHARED -DSEASTAR_API_LEVEL=6 -DSEASTAR_DEBUG -DSEASTAR_DEBUG_SHARED_PTR -DSEASTAR_DEFAULT_ALLOCATOR -DSEASTAR_SCHEDULING_GROUPS_COUNT=16 -DSEASTAR_SHUFFLE_TASK_QUEUE -DSEASTAR_TYPE_ERASE_MORE -DXXH_PRIVATE_API -DZSTD_STATIC_LINKING_ONLY -I/home/kefu/dev/redpanda/src/v -I/home/kefu/dev/redpanda/build/src/v -isystem /home/kefu/dev/redpanda/build/deps_install/include -isystem /home/kefu/dev/redpanda/build/deps_install/include/hdr -fPIC -fsanitize=undefined -fsanitize=address -fcolor-diagnostics -g -fPIC -Wall -Wextra -Werror -Wno-missing-field-initializers -std=c++20 -U_FORTIFY_SOURCE -DSEASTAR_SSTRING -Werror=unused-result "-Wno-error=#warnings" -fstack-clash-protection -fsanitize=address -fsanitize=undefined -fno-sanitize=vptr -std=c++20 -MD -MT src/v/pandaproxy/CMakeFiles/v_pandaproxy_common.dir/server.cc.o -MF src/v/pandaproxy/CMakeFiles/v_pandaproxy_common.dir/server.cc.o.d -o src/v/pandaproxy/CMakeFiles/v_pandaproxy_common.dir/server.cc.o -c /home/kefu/dev/redpanda/src/v/pandaproxy/server.cc
In file included from /home/kefu/dev/redpanda/src/v/pandaproxy/server.cc:17:
In file included from /home/kefu/dev/redpanda/src/v/pandaproxy/reply.h:16:
/home/kefu/dev/redpanda/src/v/pandaproxy/error.h:73:8: error: explicit specialization of 'std::is_error_code_enum<kafka::error_code>' after instantiation
struct is_error_code_enum<kafka::error_code> : true_type {};
       ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/usr/bin/../lib/gcc/x86_64-linux-gnu/12/../../../../include/c++/12/system_error:206:17: note: implicit instantiation first required here
             enable_if<is_error_code_enum<_ErrorCodeEnum>::value>::type>
                       ^
In file included from /home/kefu/dev/redpanda/src/v/pandaproxy/server.cc:17:
/home/kefu/dev/redpanda/src/v/pandaproxy/reply.h:103:16: error: no matching function for call to 'errored_body'
        return errored_body(e.error, e.what());
               ^~~~~~~~~~~~
/home/kefu/dev/redpanda/src/v/pandaproxy/reply.h:70:1: note: candidate function not viable: no known conversion from 'const kafka::error_code' to 'std::error_condition' for 1st argument
errored_body(std::error_condition ec, ss::sstring msg) {
^
/home/kefu/dev/redpanda/src/v/pandaproxy/reply.h:80:1: note: candidate function not viable: no known conversion from 'const kafka::error_code' to 'std::error_code' for 1st argument
errored_body(std::error_code ec, ss::sstring msg) {
^
2 errors generated.
